2015/05/05
商品管理模块新增规格管理，会员模块增加头像信息，订单模块，新增表log

+++ src/main/java/com/jifenke/lepluslive/product/repository/ProductSpecRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 /**
  * Created by wcg on 16/3/25.
  */
-public interface ProductSpecRepository extends JpaRepository<ProductSpec, Long> {
+public interface ProductSpecRepository extends JpaRepository<ProductSpec, Integer> {
 
   List<ProductSpec> findAllByProduct(Product product);
